requirements.txt
# Core
streamlit==1.31.0
pandas==2.1.4
numpy==1.26.3

# Machine Learning
scikit-learn==1.4.0
xgboost==2.0.3
lightgbm>=4.0.0
optuna==3.5.0
imbalanced-learn==0.12.0
shap==0.44.1
lifelines>=0.27.0  # Survival analysis (Kaplan-Meier, Cox PH)
statsmodels>=0.14.0  # Statistical models for regression analysis

# Visualization
plotly==5.18.0

# LLM & Vector Integration
ollama==0.1.6
langchain-community==0.0.16
faiss-cpu>=1.8.0
sentence-transformers==2.3.1
textblob==0.17.1

# Configuration
pyyaml==6.0.1

# Data Export
openpyxl==3.1.2
fpdf2==2.7.6

# Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-timeout==2.2.0

# Utilities
python-dateutil==2.8.2

# FastAPI Backend
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
